pyproject.toml
[project]
name = "logpose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pyyaml>=6.0.2",
    "ruff>=0.12.5",
]
